Финальная версия проекта + Allure-отчет

+++ src/test/java/com/test/order/OrderCreateTest.java
@@ -25,80 +25,51 @@ public class OrderCreateTest {
     @Before
     public void setUp() {
         userClient = new UserClient();
-        userData = getRandomUser();
-        token = userClient.createUser(userData)
-                .assertThat()
-                .statusCode(SC_OK)
-                .and()
-                .body("success", equalTo(true))
-                .extract()
-                .path("accessToken");
-        ;
-        userLogin = UserCredentials.from(userData);
         orderClient = new OrderClient();
+        userData = getRandomUser();
         orderData = getRandomOrder();
+        userClient.createUser(userData);
+        userLogin = UserCredentials.from(userData);
     }
     @After
     public void tearDown() {
-        userClient.deleteUser(token)
-                .assertThat()
-                .statusCode(SC_ACCEPTED)
-                .body("success", equalTo(true));
+        if (token != null) {
+            userClient.deleteUser(token);
+        }
     }
     @Test
     @DisplayName("Создание заказа с авторизацией")
     public void orderCreationWithAuth() {
-        token = userClient.loginUser(userLogin)
-                .assertThat()
-                .statusCode(SC_OK)
-                .body("success", equalTo(true))
-                .extract()
-                .path("accessToken");
+        token = userClient.loginUser(userLogin).extract().path("accessToken");
         orderClient.createWithAuthorization(orderData, token)
                 .assertThat()
                 .statusCode(SC_OK)
-                .and()
-                .body("success", equalTo(true))
-                .and()
-                .body("order.number", notNullValue());
+                .body("success", equalTo(true),"order.number", notNullValue());
     }
     @Test
-    @DisplayName("Создание заказа без авторизацией")
+    @DisplayName("Создание заказа без авторизации")
     public void orderCreationWithoutAuth() {
         orderClient.createWithoutAuthorization(orderData)
                 .assertThat()
                 .statusCode(SC_OK)
-                .body("success", equalTo(true))
-                .body("order.number", notNullValue());
+                .body("success", equalTo(true),"order.number", notNullValue());
+
     }
     @Test
     @DisplayName("Создание заказа без ингридиентов")
     public void orderCreationWithoutIngredients() {
-        token = userClient.loginUser(userLogin)
-                .assertThat()
-                .statusCode(SC_OK)
-                .body("success", equalTo(true))
-                .extract()
-                .path("accessToken");
-
+        token = userClient.loginUser(userLogin).extract().path("accessToken");
         orderData.setIngredients(null);
         orderClient.createWithAuthorization(orderData, token)
                 .assertThat()
                 .statusCode(SC_BAD_REQUEST)
                 .and()
-                .body("success", equalTo(false))
-                .and()
-                .body("message", equalTo("Ingredient ids must be provided"));
+                .body("success", equalTo(false),"message", equalTo("Ingredient ids must be provided"));
     }
     @Test
     @DisplayName("Создание заказа с неправильным хешем")
     public void orderCreationWithIncorrectHash() {
-        token = userClient.loginUser(userLogin)
-                .assertThat()
-                .statusCode(SC_OK)
-                .body("success", equalTo(true))
-                .extract()
-                .path("accessToken");
+        token = userClient.loginUser(userLogin).extract().path("accessToken");
         List<String> incorrectHash = new ArrayList<>();
         incorrectHash.add(RandomStringUtils.randomAlphabetic(10));
         orderData.setIngredients(incorrectHash);
