+++ src/test/java/com/test/order/OrderGetTest.java
@@ -2,7 +2,10 @@ package com.test.order;
 
 import com.example.client.*;
 import com.example.data.*;
+import io.qameta.allure.Step;
 import io.qameta.allure.junit4.DisplayName;
+import org.junit.After;
+import org.junit.Before;
 import org.junit.Test;
 
 import static com.example.data.UserGenerator.getRandomUser;
@@ -13,51 +16,37 @@ import static org.hamcrest.CoreMatchers.*;
 public class OrderGetTest {
     private UserClient userClient;
     private OrderClient orderClient;
-    private UserData user;
-    private UserCredentials userLogin;
-    private OrderData order;
+    private UserData userData;
     private String token;
-
+    @Before
+    @Step("Создание тестовых данных")
+    public void setUp() {
+        userClient = new UserClient();
+        orderClient = new OrderClient();
+        userData = getRandomUser();
+        token = userClient.createUser(userData).extract().path("accessToken");
+    }
+    @After
+    @Step("Удаление тестовых данных")
+    public void cleanUp() {
+        if(token!=null) {
+            userClient.deleteUser(token);
+        }
+    }
     @Test
     @DisplayName("Получение заказа с авторизацией")
     public void getOrdersWithAuth() {
-        userClient = new UserClient();
-        user = getRandomUser();
-        token = userClient.createUser(user)
-                .assertThat()
-                .statusCode(SC_OK)
-                .body("success", equalTo(true))
-                .extract()
-                .path("accessToken");
-        ;
-        userLogin = UserCredentials.from(user);
-        token = userClient.loginUser(userLogin)
-                .assertThat()
-                .statusCode(SC_OK)
-                .body("success", equalTo(true))
-                .extract()
-                .path("accessToken");
-
-        orderClient = new OrderClient();
         orderClient.getWithAuthorization(token)
                 .assertThat()
                 .statusCode(SC_OK)
                 .body("success", equalTo(true));
-
-        userClient.deleteUser(token)
-                .assertThat()
-                .statusCode(SC_ACCEPTED)
-                .body("success", equalTo(true));
     }
-
     @Test
-    @DisplayName("Получение заказа без авторизацией")
+    @DisplayName("Получение заказа без авторизации")
     public void getOrdersWithoutAuth() {
-        orderClient = new OrderClient();
         orderClient.getWithoutAuthorization()
                 .assertThat()
                 .statusCode(SC_UNAUTHORIZED)
-                .body("success", equalTo(false))
-                .body("message", equalTo("You should be authorised"));
+                .body("success", equalTo(false),"message", equalTo("You should be authorised"));
     }
 }
